#### CrossApp 1.2.0 Beta 更新内容：
#####1.简化适配概念
1.删除CCRect、CCPoint、CCSize、CADipRect、CADipPoint、CADipSize
2.增加新的DRect、DPoint、DSize，以dp为单位。(在PPI为320的设备上，1px=1dp)

#####2.支持C++11

#####3.新增视频播放器
1.CAVideoPlayerView
2.CAVideoPlayerControlView

#####4.支持Emoji字体
1.CALabel支持显示系统默认Emoji字体
2.CATextField支持输入系统键盘Emoji表情
2.CATextView支持输入系统键盘Emoji表情

#####5.优化文字系统
兼容iOS9系统字体PingFang.ttc

#####6.删除CCActions系统
请使用CAViewAnimation系统

#####7.删除CCArray、CCDictionary
请使用CAVector、CADeque、CAList、CAMap

#####8.删除CCString
请使用std::string

#####9.优化PC模拟器
将之前PC端模拟器与系统像素比1：1改为2：1，以满足小屏幕有足够空间显示模拟器

#####10.部分控件增加PC模式
例如，CAScrollView在PC端可开启PC模式，以支持PC操作方式

#####11.iOS版本兼容修改
新版放弃支持iOS6.0+，最低版本提升到iOS7.0

#####12.新增Test-HD
平板模式的test

#####13.其他一些细节的优化

+++ CrossApp/platform/CAFreeTypeFont.cpp
@@ -327,7 +327,7 @@ int CAFreeTypeFont::cutStringByWidth(const std::string& text, int iLimitWidth, i
     return nCharPos;
 }
 
-int CAFreeTypeFont::cutStringByDSize(std::string& text, const DSize& lableSize, const std::string& pFontName, int nFontSize, bool bWordWrap, int iLineSpacing, bool bBold, bool bItalics)
+int CAFreeTypeFont::cutStringByDSize(std::string& text, const DSize& lableSize, const std::string& pFontName, unsigned long nFontSize, bool bWordWrap, int iLineSpacing, bool bBold, bool bItalics)
 {
 	std::u32string cszTemp;
 	std::string cszNewText = text;
@@ -500,7 +500,7 @@ void  CAFreeTypeFont::drawText(FTLineInfo* pInfo, unsigned char* pBuffer, FT_Vec
         FT_Glyph image = glyph->image;
 		if (image == NULL)
 		{
-            CAImage* pEmoji = CAEmojiFont::getInstance()->getEmojiImage(glyph->c, m_inFontSize);
+            CAImage* pEmoji = CAEmojiFont::getInstance()->getEmojiImage((unsigned int)glyph->c, m_inFontSize);
             if (pEmoji)
             {
                 pEmoji = CAImage::scaleToNewImageWithImage(pEmoji, DSize(m_inFontSize, m_inFontSize));
